Add test suite for prime_function
We separate out the prime_function module tests into a test suite
which is run using nose. Specifically, the user runs the suite
using

$ sage runtests.py

provided.

The purpose of this separation is to isolate where the testing
examples occur as well as to make sure that the tests for
prime_function run independently of the other code in the library.
(e.g. forward_problem)

Other modules should be tested in this way.